task_2/myproject/requirements.txt
dj-database-url==2.2.0
Django==5.0.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9